haciendo funcionar selectAll notificaciones

+++ src/pages/notification/notification.ts
@@ -31,6 +31,9 @@ export class NotificationPage {
   filter: any = {name: '', location: '', hasResult: true};
   notifications: Array<{user: string, message: string, time: string, viewed: boolean, isChecked: boolean}> = [];
   deleteAll: boolean = false;
+  deleteAllChange: boolean = false;
+  notDeleteAllChange: boolean = false;
+  selectedNotifications: Number;
   
   constructor(public navCtrl: NavController, public navParams: NavParams, public confData: NotificationsData) {  }
 
@@ -42,19 +45,50 @@ export class NotificationPage {
     });
   }
 
+  refreshNotifications(){
+    this.confData.refreshNotifications().subscribe((data: any) => {
+      this.UnselectAll();
+      this.notifications = data;
+    });
+  }
+
   selectAll(){
+    if(!this.deleteAllChange){
+      this.notDeleteAllChange = true;
+      this.selectedNotifications = this.deleteAll? 0: this.notifications.length;
       for (let i = this.notifications.length - 1; i >= 0; i--) {
         this.notifications[i].isChecked = this.deleteAll;
       }
+    }
+    this.deleteAllChange = false;
+  }
+
+  UnselectAll(){
+    this.deleteAll = false;
+    this.selectedNotifications = 0;
+    console.log('deleteAll: '+this.deleteAll);
+  }
+
+  toogleNotification(notification){
+    if(!this.notDeleteAllChange){
+      this.deleteAllChange = true;
+      if (notification.isChecked) {
+        this.selectedNotifications++;
+      } else {
+        this.selectedNotifications--;
+      }
+      this.deleteAll = this.selectedNotifications == this.notifications.length;
+      console.log('deleteAll: '+this.selectedNotifications +"=="+ this.notifications.length);
+    }
   }
 
   dimissNotifications(){
     	for (let i = this.notifications.length - 1; i >= 0; i--) {
     		if(this.notifications[i].isChecked){
           this.confData.deleteNotification(this.notifications[i]);
-    			//this.notifications.splice(i, 1);
     		}
     	}
+      this.UnselectAll();
   }
 
   checkNotifications(){
@@ -63,10 +97,12 @@ export class NotificationPage {
           this.notifications[i].viewed = true;
         }
       }
+      this.UnselectAll();
   }
 
   updateNotifications(){
     this.filter.name =this.queryText;
+    this.UnselectAll();
   }
